+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This change log follows the [Keep a Changelog standards](http://keepachangelog.c
 * WP-API integration
 	* This requires an upgrade to WordPress 4.4+ or installation of rest-api plugin
 * Plugin architecture now built on [Jaxion] framework
+* Prism plugins:
+	* Show invisibles: Display tabs and line returns as characters
 
 #### Changed ####
 * Improve JavaScript architecture